+++ lib/kagayaku_modules.dart
@@ -2,5 +2,6 @@ library;
 
 export 'src/kagayaku_modules_base.dart';
 export 'src/models/novel.dart';
+export 'src/models/module_info.dart';
 export 'src/utils/get_source.dart';
 export 'src/utils/languages.dart';