+++ CHANGELOG.md
@@ -10,4 +10,11 @@
 
 ### Enhancements
 - Changed package imports to relative imports for better modularity.
-- http changed per dio for http requests
+- http changed per dio for http requests.
+
+## 1.0.2
+
+### Enhancements
+- Changed the way to get source url.
+- Returned `null` if the language is not supported.
+- Added `ModuleInfo` class to store module information.